CodeRefactor: - feat: Created TrajectorySolver base class - feat: Implemented base class in approaches class
